Block editor: hooks: share block settings
